fixup! Feat(variables-scss): Export to javascript #DS-1437
