+++ exporters/variables-scss/src/generators/__tests__/fileGenerator.test.ts
@@ -2,15 +2,16 @@ import fs from 'fs';
 import path from 'path';
 import { Supernova, Token, TokenGroup, TokenTheme } from '@supernovaio/sdk-exporters';
 import { generateFiles, generateIndexFile, generateOutputFilesByThemes } from '../fileGenerator';
-import { exampleMockedGroups, exampleMockedTokens } from '../../../tests/fixtures/mockedExampleTokens';
+import { exampleDimensionAndStringTokens } from '../../../tests/fixtures/exampleDimensionAndStringTokens';
 import { nonThemedFilesData } from '../../config/fileConfig';
+import { exampleGroups } from '../../../tests/fixtures/exampleGroups';
 
 const mockedExpectedResult = fs.readFileSync(
   path.join(__dirname, '../../../tests/fixtures/exampleFileContent.scss'),
   'utf-8',
 );
 const mappedTokens: Map<string, Token> = new Map([]);
-const tokenGroups: Array<TokenGroup> = exampleMockedGroups;
+const tokenGroups: Array<TokenGroup> = exampleGroups;
 const emptyFile = `/* This file was generated by Supernova, don't change manually */\n\n`;
 const indexFile = `@forward 'borders';
 @forward 'other';
@@ -44,7 +45,7 @@ columns: gridColumns,
 describe('fileGenerator', () => {
   describe('generateOutputFilesByThemes', () => {
     it('should generate output files by themes', async () => {
-      const tokens = Array.from(exampleMockedTokens.values());
+      const tokens = Array.from(exampleDimensionAndStringTokens.values());
       const sdk = {
         tokens: {
           computeTokensByApplyingThemes: jest.fn().mockResolvedValue(tokens),
@@ -93,7 +94,7 @@ describe('fileGenerator', () => {
 
   describe('generateFiles', () => {
     it('should generate files', () => {
-      const tokens = Array.from(exampleMockedTokens.values());
+      const tokens = Array.from(exampleDimensionAndStringTokens.values());
       const files = generateFiles(tokens, mappedTokens, tokenGroups, nonThemedFilesData, false);
 
       expect(files).toStrictEqual([
